Update binary mixture experiment to run properly
This ensures the binary mixture experiment runs appropriately on Quartz.
This does include a workaround for an issue with setting up an
experiment with statdyn-simulation, in not running the creation with
MPI. Additionally I have changed the number of cores to 8 since allowing
this to run on the older nodes of quartz.

A significant change that actually changes the function of this
configuration is creating the setup parameters for the pbs script which
hadn't previously been configured correctly.