Include notebook for studying stoke-einstein
This is the notebook where I have been implementing much of the
optimisations for the analysis of the breakdown of Stokes-Einstein
dynamics. There is still much to understand about this behaviour so
there is a lot of code in this notebook which will eventually make it's
way into python packages for simpler analysis.