Development of for understanding relaxation
The motions that we are generating are expeced to be brownian, i.e. at
high T these are the expected distribution of motions and the deviations
from Brownian are the interesting behaviour in dynamic inhomogeneities.

I have created some simulations of Brownian motion and from this
calculated the same dynamic properties that I am observing as a
reference for the expected behaviour of these new quantities.