Create script to compute relaxations
This takes the computation of relaxation qualities which was taking
place in one of the notebooks and turns it into a python script. This
makes the long running analysis simpler to run. As part of creating a
script to run the analysis I have also created a makefile rule which
will run the analysis with the command `make relaxations`.