Calculate both mean and harmonic mean of molecular_relaxations
There are a number of quantities for the molecular_relaxations where I
require both the mean and the harmonic mean. This commit adds the
calculation of the harmonic mean to all the quantities.

This also changes the name of the columns in the dataframe, having the
suffix `_mean` for the mean values and `_hmean` for the harmonic mean.