Fix calculation of max_relaxation to not include nan
The calculation of the relaxation time by the max_relaxation function
was giving anomalous values, a result of having NaN values in the array,
which are considered the largest. This swaps the function `np.argmax`
with `np.nanargmax` which will ignore the NaN values in the array when
finding the largest value. This commit includes the addition of tests to
ensure the behaviour remains as expected.