Add notebook to generate all figures for publication
The figures notebook will generate all the figures for the current
version of the dynamics paper (v12). Figures are output to the figures
directory labelled by the figure reference from the paper.

While the figures are in a notebook, they can also all be generated
using the command `make figures` which will run the requisite commands.