EXP Fix bugs in experiment file
The experiment file now performs the experiment I would like it to
perform, iterating over a different range of temperatures for the high
and low pressure simulations.

Additionally it makes complete use of the requires and creates
arguments, which now also have the pressure and temperature arranged to
be consistent with the rest of the simulation output.

Additionally this includes a workaround for there being no dynamics
output, i.e. the trajectory-* files.