EXP Double simulation times for dynamics
A number of the dynamics simulations didn't reach equilibrium during the
initial simulation runs, which was solved by taking the configurations
from the end of the production run and collecting data again. This
commits effectively adds that additional step to the experiment by
doubling all the simulation lengths, apart from the lowest temperature
which is quadrupled since those simulations were still not equilibrated
after the extended time.