enh: Calculate the dynamics using a wave number
This uses the wave-number for computing the dynamics quantities. Here I
am computing the quantities using the wave-number as a reference length.
There is a different wave-number for each pressure, representing the
different values for each pressure, reflecting the differences in
distances for the pressures.

BREAKING CHANGE

This changes the calculations of the dynamics quantities being the
distance we are looking at for various dynamics quantities.