feat: Separate processing of high and low pressure trajectories
The high and low pressures have different characteristic distances, that
is the wave-number. To treat these trajectories properly, they have to
be processed separately. While there is a different value of the
wave-number for each pressure, due to the large difference between the
pressures, at the range of temperatures studied there is little change
in the wave number, so only a single value is used.