processing: Create script to collate datasets
Rather than putting all the files into a single dataset, which means
that when one breaks, all the files have to be recomputed, I have now
transitioned to a single dataset for each trajectory. This means that
there needs to be an additional step of collating all the individual
files into a single file for ease of use.

This creates the script for collating all the single files into one
large file.