notebook: Update notebooks, ensuring they work again
The one exception to this is notebook 20, which is the publication
figures notebook which is very outdated and requires a lot of work to
update. Since I am not using the figures from this currently, I am just
leaving this for now. The rest of the notebooks have been updated to
work and have been confirmed to work.