fix: Use terminology inv_temp_norm
The normalised temperature is T/T_m, providing a method of comparing
temperatures across substances with different melting points. For the
dynamics, the value we are more interested in is 1/T, that is the
inverse temperature, so we take the inverse of the normalised
temperature being T_m / T.

To make this comparison clear, I am using the term temp_norm to refer
to the normalised temperature, and inv_temp_norm to refer to it's
inverse.