analysis: Only calculate msr for small rotations
At large times I have issues with the MSR in that the rotations could
have been going in either direction. This makes some of the larger
values of the MSR less reliable and accurate. As a way to more
accurately calculate the rotational diffusion constant I am only using
values up to 10.

This removes the values above 10 by masking them with values of 1000.
I would prefer to use nan, however the analysis tool in sdanalysis
requires proper floating point values. As a way around this I have set
all the values above 10 to 1000 which is ignored by the sdanalysis code
giving the desired result.

Ref https://github.com/malramsay64/statdyn-analysis/issues/158