analysis: Create function to fit VTF to relaxations
This creates a function to fit the VTF equations to the dynamics
results. This also includes a flag for adding the resulting fitted
equation to the figure of the relaxation.

This enables #38, #17 and #25, all of which require the results of the
fitting either in the figures or for further analysis.